[javadocs] Fill in using 'jautodoc' all test javadocs
reason 1: to get rid of maven warnings
reason 2: regardless its helpful in so many cases.  the 'jautodoc' does a great job, many disagree here but hello javadoc plugin complains if this is not done.  So better have a baseline then nothing at all.  Then if more is to be put there as for example what it does, then add it, its stubbed and ready for that.
reason 3: see reason 1