[packages] modular requires packages match on source / test side, fixing so we have placeholders
Tests are under a folder test that are unrelated to the source side, javadocs try to force modular and will not work if they do not include proper packaging so added these.  They serve no other purpose other than to fix the site generation due to release failing when javadocs cannot work here.